requirements.txt
APScheduler==3.9.1
diff-match-patch==20181111
arabic-reshaper==2.1.3
asgiref==3.6.0
Babel==2.9.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==36.0.2
cssselect2==0.5.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2.1
django-allauth==0.50.0
django-apscheduler==0.6.2
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-import-export==2.1.0
django-daraja==1.2.0
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==6.1.0
git+https://github.com/9409010617086/django-pwa.git
django-slick-reporting==0.6.3
django-summernote==0.8.20.0
django-user-visit==0.5.1
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
lxml==4.8.0
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.45
phonenumberslite==8.12.45
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-9409010617086==0.1.0.post0
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.4
stripe==2.69.0
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.63.1
tzdata==2023.3
tzlocal==4.1
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.6
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
backports.zoneinfo==0.2.1
django-credit-cards==0.4.1
